split feature in multiple file
