+++ src/features/Newsletters/interfaces.ts
@@ -1,0 +1,13 @@
+export interface INewsLetter {
+  id: string;
+  image: string;
+  description: string;
+  title: string;
+  site: string;
+  subscriptions: string[];
+}
+
+export interface PageState {
+  currentUser: string;
+  updateCurrentUser: (user: string) => void;
+}
